requirements.txt
fastapi
uvicorn[standard]
asyncpg
chromadb
sentence-transformers
redis
google-adk
python-dotenv
pydantic-settings